requirements.txt
transformers==4.12.3

multiprocess==0.70.9
numpy==1.17.2
pyrouge==0.1.3
pytorch-transformers==1.2.0
tensorboardX==1.9
torch==1.10.0
bert-extractive-summarizer 0.8.1
